requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
attrs==25.4.0
black==25.12.0
blosc2==3.12.2
botocore==1.34.54
brotli==1.2.0
certifi==2022.12.7
charset-normalizer==3.4.4
click==8.3.1
cmake==3.25.0
contourpy==1.3.3
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
fastapi==0.124.2
ffmpy==1.0.0
filelock==3.20.0
flake8==7.0.0
flake8-bugbear==24.2.6
flake8-comprehensions==3.14.0
fonttools==4.61.0
fsspec==2025.12.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
gitpython==3.1.45
gradio==6.1.0
gradio-client==2.0.1
groovy==0.1.2
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.2.2
idna==3.11
imageio==2.37.2
imath==0.0.2
iopath==0.1.10
isort==7.0.0
jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
lit==15.0.7
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
mypy-extensions==1.1.0
ndindex==1.10.1
networkx==3.6.1
ninja==1.13.0
numexpr==2.14.1
numpy==1.26.4
opencv-python==4.11.0.86
openexr==3.4.4
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.5.1
portalocker==3.2.0
protobuf==4.25.3
psutil==7.1.3
py-cpuinfo==9.0.0
pycodestyle==2.11.1
pydantic==2.12.4
pydantic-core==2.41.5
pydub==0.25.1
pyflakes==3.2.0
pygments==2.19.2
pyparsing==3.2.5
pyre-extensions==0.0.29
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
rich==14.2.0
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.16.3
semantic-version==2.10.0
sentry-sdk==2.47.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
starlette==0.50.0
sympy==1.14.0
tables==3.10.2
tabulate==0.9.0
termcolor==3.2.0
timm==1.0.22
tomlkit==0.13.3
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
tqdm==4.67.1
triton==2.0.0
typer==0.20.0
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2024.1
urllib3==1.26.13
uvicorn==0.38.0
wandb==0.23.1
xformers==0.0.20
yacs==0.1.8